Move JobTable to its own file
